refactor(tmux): split client.go into 5 focused files
Split 501-line monolithic client.go into 5 single-responsibility files:
- context.go: context/pane validation and status operations (111 lines)
- navigation.go: jump/navigate operations (131 lines)
- lists.go: list sessions/windows/panes operations (102 lines)
- environment.go: environment variable operations (72 lines)
- client.go: reduced to 119 lines with core types, interface, and Run method

Additional changes:
- Remove client.go from linter exclude list in scripts/lint.sh
- No breaking changes to public API
- All tests pass (24/24)
- Linter passes on all files

Follows existing pattern of config.go and errors.go in the package.
Improves maintainability and code organization.
