+++ internal/tmux/client.go
@@ -7,10 +7,8 @@ import (
 	"context"
 	"fmt"
 	"os/exec"
-	"strings"
 	"time"
 
-	"github.com/cristianoliveira/tmux-intray/internal/colors"
 	"github.com/cristianoliveira/tmux-intray/internal/errors"
 )
 
@@ -108,386 +106,6 @@ func (c *DefaultClient) runCommand(args ...string) (string, string, error) {
 
 	err := cmd.Run()
 	return stdout.String(), stderr.String(), err
-}
-
-// GetCurrentContext returns the current tmux session/window/pane context.
-func (c *DefaultClient) GetCurrentContext() (TmuxContext, error) {
-	format := "#{session_id} #{window_id} #{pane_id} #{pane_pid}"
-	stdout, stderr, err := c.Run("display", "-p", format)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return TmuxContext{}, fmt.Errorf("failed to get tmux context: %w", err)
-	}
-
-	// Trim whitespace from the output
-	stdout = strings.TrimSpace(stdout)
-
-	// Split by space - tmux format string produces 4 parts
-	// Format: #{session_id} #{window_id} #{pane_id} #{pane_pid}
-	parts := strings.Split(stdout, " ")
-
-	// Filter out empty strings that might result from multiple spaces
-	var filteredParts []string
-	for _, part := range parts {
-		if part != "" {
-			filteredParts = append(filteredParts, part)
-		}
-	}
-
-	// Validate that we have exactly 4 parts
-	if len(filteredParts) != 4 {
-		return TmuxContext{}, fmt.Errorf("unexpected format - expected 4 parts, got %d", len(filteredParts))
-	}
-
-	// Validate that captured context values are non-empty
-	if filteredParts[0] == "" || filteredParts[1] == "" || filteredParts[2] == "" {
-		return TmuxContext{}, fmt.Errorf("invalid context - session_id, window_id, or pane_id is empty")
-	}
-
-	ctx := TmuxContext{
-		SessionID: filteredParts[0],
-		WindowID:  filteredParts[1],
-		PaneID:    filteredParts[2],
-		PanePID:   filteredParts[3],
-	}
-	return ctx, nil
-}
-
-// ValidatePaneExists checks if a pane exists in a given session and window.
-func (c *DefaultClient) ValidatePaneExists(sessionID, windowID, paneID string) (bool, error) {
-	target := sessionID + ":" + windowID
-	stdout, stderr, err := c.Run("list-panes", "-t", target, "-F", "#{pane_id}")
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return false, fmt.Errorf("failed to list panes: %w", err)
-	}
-
-	// Each pane ID is on a separate line
-	lines := strings.Split(strings.TrimSpace(stdout), "\n")
-	// Trim paneID input in case it has whitespace
-	paneID = strings.TrimSpace(paneID)
-	for _, line := range lines {
-		// Trim each pane ID from output in case of trailing whitespace
-		if strings.TrimSpace(line) == paneID {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-// jumpToPane is the private implementation that accepts a custom error handler.
-func (c *DefaultClient) jumpToPane(sessionID, windowID, paneID string, handler errors.ErrorHandler) (bool, error) {
-	// Validate required fields (sessionID and windowID)
-	if sessionID == "" || windowID == "" {
-		return false, ErrInvalidTarget
-	}
-
-	currentCtx, err := c.GetCurrentContext()
-	if err != nil {
-		colors.Debug("JumpToPane: failed to get current tmux context: " + err.Error())
-	} else if currentCtx.SessionID != sessionID {
-		colors.Debug(fmt.Sprintf("JumpToPane: switching client to session %s", sessionID))
-		_, stderr, err := c.Run("switch-client", "-t", sessionID)
-		if err != nil {
-			if stderr != "" {
-				colors.Debug("JumpToPane: stderr: " + stderr)
-			}
-			return false, fmt.Errorf("switch client to session %s: %w", sessionID, err)
-		}
-	}
-
-	// paneID is optional - if empty, jump to window only
-	if paneID == "" {
-		return c.jumpToWindowOnly(sessionID, windowID)
-	}
-
-	return c.jumpToPaneWithValidation(sessionID, windowID, paneID, handler)
-}
-
-// jumpToWindowOnly switches to the specified session and selects the window.
-func (c *DefaultClient) jumpToWindowOnly(sessionID, windowID string) (bool, error) {
-	colors.Debug(fmt.Sprintf("JumpToPane: switching client to session %s (window-only jump)", sessionID))
-	_, stderr, err := c.Run("switch-client", "-t", sessionID)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return false, fmt.Errorf("switch client to session %s: %w", sessionID, err)
-	}
-
-	targetWindow := sessionID + ":" + windowID
-	colors.Debug(fmt.Sprintf("JumpToPane: selecting window %s (window-only jump)", targetWindow))
-	_, stderr, err = c.Run("select-window", "-t", targetWindow)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("JumpToPane: stderr: " + stderr)
-		}
-		return false, fmt.Errorf("window %s does not exist: %w", targetWindow, err)
-	}
-	colors.Debug(fmt.Sprintf("JumpToPane: successfully selected window %s", targetWindow))
-	return true, nil
-}
-
-// jumpToPaneWithValidation switches to the specified session, window, and pane after validation.
-func (c *DefaultClient) jumpToPaneWithValidation(sessionID, windowID, paneID string, handler errors.ErrorHandler) (bool, error) {
-	paneExists, err := c.ValidatePaneExists(sessionID, windowID, paneID)
-	if err != nil {
-		return false, fmt.Errorf("pane validation failed: %w", err)
-	}
-
-	colors.Debug(fmt.Sprintf("JumpToPane: pane validation for %s:%s - exists: %v", sessionID, windowID, paneExists))
-
-	// Switch client to target session before selecting window/pane
-	colors.Debug(fmt.Sprintf("JumpToPane: switching client to session %s", sessionID))
-	_, stderr, err := c.Run("switch-client", "-t", sessionID)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return false, fmt.Errorf("switch client to session %s: %w", sessionID, err)
-	}
-
-	// Select the window first
-	targetWindow := sessionID + ":" + windowID
-	colors.Debug(fmt.Sprintf("JumpToPane: selecting window %s", targetWindow))
-	_, stderr, err = c.Run("select-window", "-t", targetWindow)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("JumpToPane: stderr: " + stderr)
-		}
-		return false, fmt.Errorf("window %s does not exist: %w", targetWindow, err)
-	}
-
-	// If pane doesn't exist, show warning and fall back to window selection
-	if !paneExists {
-		handler.Warning("Pane " + paneID + " does not exist in window " + targetWindow + ", jumping to window instead")
-		colors.Debug(fmt.Sprintf("JumpToPane: falling back to window selection (pane %s not found)", paneID))
-		return true, nil
-	}
-
-	// Pane exists, select it using the correct tmux pane syntax: "sessionID:windowID.paneID"
-	targetPane := sessionID + ":" + windowID + "." + paneID
-	colors.Debug(fmt.Sprintf("JumpToPane: selecting pane %s", targetPane))
-	_, stderr, err = c.Run("select-pane", "-t", targetPane)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("JumpToPane: stderr: " + stderr)
-		}
-		return false, fmt.Errorf("failed to select pane %s: %w", targetPane, err)
-	}
-
-	colors.Debug(fmt.Sprintf("JumpToPane: successfully selected pane %s", targetPane))
-	return true, nil
-}
-
-// JumpToPane jumps to the specified pane or window.
-// If paneID is empty, jumps to the window only.
-// Returns true if jump succeeded, false if failed.
-// Preconditions: sessionID and windowID must be non-empty; paneID is optional.
-func (c *DefaultClient) JumpToPane(sessionID, windowID, paneID string) (bool, error) {
-	return c.jumpToPane(sessionID, windowID, paneID, defaultCLIHandler)
-}
-
-// JumpToPaneWithHandler jumps to the specified pane or window with a custom error handler.
-// This allows TUI and other UIs to handle errors differently than the CLI.
-// If paneID is empty, jumps to the window only.
-// Returns true if jump succeeded, false if failed.
-// Preconditions: sessionID and windowID must be non-empty; paneID is optional.
-func (c *DefaultClient) JumpToPaneWithHandler(sessionID, windowID, paneID string, handler errors.ErrorHandler) (bool, error) {
-	return c.jumpToPane(sessionID, windowID, paneID, handler)
-}
-
-// SetEnvironment sets a tmux environment variable.
-func (c *DefaultClient) SetEnvironment(name, value string) error {
-	_, stderr, err := c.Run("set-environment", "-g", name, value)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return fmt.Errorf("failed to set environment variable %s: %w", name, err)
-	}
-	return nil
-}
-
-// GetEnvironment gets a tmux environment variable value.
-func (c *DefaultClient) GetEnvironment(name string) (string, error) {
-	stdout, stderr, err := c.Run("show-environment", "-g", name)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return "", fmt.Errorf("failed to get environment variable %s: %w", name, err)
-	}
-
-	// Output could be "NAME=value" or empty line
-	lines := strings.Split(strings.TrimSpace(stdout), "\n")
-	for _, line := range lines {
-		if strings.HasPrefix(line, name+"=") {
-			return strings.TrimPrefix(line, name+"="), nil
-		}
-	}
-	return "", fmt.Errorf("environment variable %s not found", name)
-}
-
-// HasSession checks if tmux server is running.
-func (c *DefaultClient) HasSession() (bool, error) {
-	_, stderr, err := c.Run("has-session")
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return false, ErrTmuxNotRunning
-	}
-	return true, nil
-}
-
-// SetStatusOption sets a tmux status option.
-func (c *DefaultClient) SetStatusOption(name, value string) error {
-	// First check if tmux is running
-	running, err := c.HasSession()
-	if err != nil {
-		return err
-	}
-	if !running {
-		return ErrTmuxNotRunning
-	}
-
-	_, stderr, err := c.Run("set", "-g", name, value)
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return fmt.Errorf("failed to set status option %s: %w", name, err)
-	}
-	return nil
-}
-
-// ListSessions returns all tmux sessions as a map of session ID to name.
-func (c *DefaultClient) ListSessions() (map[string]string, error) {
-	stdout, stderr, err := c.Run("list-sessions", "-F", "#{session_id}\t#{session_name}")
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return nil, fmt.Errorf("failed to list sessions: %w", err)
-	}
-
-	sessions := make(map[string]string)
-	lines := strings.Split(strings.TrimSpace(stdout), "\n")
-	for _, line := range lines {
-		if line == "" {
-			continue
-		}
-		parts := strings.SplitN(line, "\t", 2)
-		if len(parts) == 2 {
-			sessions[parts[0]] = parts[1]
-		}
-	}
-	return sessions, nil
-}
-
-// GetSessionName returns the name of a session by its ID.
-func (c *DefaultClient) GetSessionName(sessionID string) (string, error) {
-	stdout, stderr, err := c.Run("display-message", "-t", sessionID, "-p", "#S")
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return "", fmt.Errorf("get session name: %w", err)
-	}
-
-	// Trim whitespace from the output
-	sessionName := strings.TrimSpace(stdout)
-
-	// Validate that session name is non-empty
-	if sessionName == "" {
-		return "", ErrSessionNotFound
-	}
-
-	return sessionName, nil
-}
-
-// ListWindows returns all tmux windows as a map of window ID to name.
-func (c *DefaultClient) ListWindows() (map[string]string, error) {
-	stdout, stderr, err := c.Run("list-windows", "-a", "-F", "#{window_id}\t#{window_name}")
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return nil, fmt.Errorf("failed to list windows: %w", err)
-	}
-
-	windows := make(map[string]string)
-	lines := strings.Split(strings.TrimSpace(stdout), "\n")
-	for _, line := range lines {
-		if line == "" {
-			continue
-		}
-		parts := strings.SplitN(line, "\t", 2)
-		if len(parts) == 2 {
-			windows[parts[0]] = parts[1]
-		}
-	}
-	return windows, nil
-}
-
-// ListPanes returns all tmux panes as a map of pane ID to name.
-func (c *DefaultClient) ListPanes() (map[string]string, error) {
-	stdout, stderr, err := c.Run("list-panes", "-a", "-F", "#{pane_id}\t#{pane_title}")
-	if err != nil {
-		if stderr != "" {
-			colors.Debug("stderr: " + stderr)
-		}
-		return nil, fmt.Errorf("failed to list panes: %w", err)
-	}
-
-	panes := make(map[string]string)
-	lines := strings.Split(strings.TrimSpace(stdout), "\n")
-	for _, line := range lines {
-		if line == "" {
-			continue
-		}
-		parts := strings.SplitN(line, "\t", 2)
-		if len(parts) == 2 {
-			panes[parts[0]] = parts[1]
-		}
-	}
-	return panes, nil
-}
-
-// GetTmuxVisibility gets the tmux visibility state from environment variable.
-func (c *DefaultClient) GetTmuxVisibility() (bool, error) {
-	value, err := c.GetEnvironment("TMUX_INTRAY_VISIBLE")
-	if err != nil {
-		// Environment variable not set is not an error - return false
-		if strings.Contains(err.Error(), "not found") {
-			return false, nil
-		}
-		return false, fmt.Errorf("get tmux visibility: %w", err)
-	}
-
-	// Parse boolean value from string
-	return value == "true", nil
-}
-
-// SetTmuxVisibility sets the tmux visibility state via environment variable.
-func (c *DefaultClient) SetTmuxVisibility(visible bool) error {
-	var value string
-	if visible {
-		value = "true"
-	} else {
-		value = "false"
-	}
-
-	if err := c.SetEnvironment("TMUX_INTRAY_VISIBLE", value); err != nil {
-		return fmt.Errorf("set tmux visibility: %w", err)
-	}
-
-	return nil
 }
 
 // Run executes a tmux command with the given arguments.
