+++ scripts/lint.sh
@@ -20,7 +20,6 @@ check_file_length() {
     local exclude_files=(
         "./internal/tui/state/model_keys.go"
         "./cmd/tmux-intray/status-panel.go"
-        "./internal/tmux/client.go"
     )
     while IFS= read -r -d '' file; do
         # Skip grandfathered large files
